Update README.md
Aspose.Email for .NET v24.9.0

Aspose.Email for .NET is a comprehensive set of Email Processing APIs that can be used to build cross-platform applications with the ability to create mails, manipulate, analyze, convert and transmit messages without using Microsoft Outlook or Office Automation. The API provides the email manipulation features such as addition, extraction or removal of attachments from a message object, customization of message headers by adding or removing recipients or changing the subject and more. The overview of these features and their use with code samples you can find in our documentation.

[Download Aspose.Email for .NET v24.9.0](https://developer.team/dotnet/35023-asposeemail-for-net-v2490.html)